mini-contract-api/requirements.txt
# Web 框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# 数据库
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
alembic>=1.13.0

# 缓存
redis[hiredis]>=5.0.0

# 认证与安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP 客户端
httpx>=0.27.0

# 微信 SDK
wechatpy>=2.1.0

# 文件存储
python-minio>=7.2.0

# 数据校验
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 环境变量
python-dotenv>=1.0.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
respx>=0.21.0
httpx>=0.27.0
